refactor: types/interfaces closer to their usage for better locality of behavior
